# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoList API",
    "version": "1.0.0",
    "description": "인증된 사용자가 할 일을 카테고리와 일정(시작일/종료일) 기반으로 관리할 수 있는 TodoList 웹 애플리케이션 API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "로컬 개발 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 인증 토큰. 로그인(POST /api/auth/login) 후 발급받은 토큰을 입력합니다."
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "description": "할 일 정보. status 필드는 DB에 저장되지 않으며 is_completed, start_date, end_date로 런타임 계산됨",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "할 일 고유 식별자"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "할 일 소유자 ID (BR-02, BR-05)"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "분류 카테고리 ID (미지정 시 기본 카테고리 자동 배정 — BR-03)"
          },
          "title": {
            "type": "string",
            "description": "할 일 제목"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "할 일 상세 설명 (선택값)"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "시작일 (선택값, BR-07)"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "종료일 (선택값, BR-07). startDate 이후여야 함 (BR-06)"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "완료 여부"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "등록일시"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "수정일시"
          }
        },
        "required": ["id", "userId", "categoryId", "title", "isCompleted", "createdAt", "updatedAt"]
      },
...
      "CreateTodoRequest": {
        "type": "object",
        "description": "할 일 등록 요청 바디 (UC-04)",
        "properties": {
          "title": {
            "type": "string",
            "description": "할 일 제목 (필수)"
          },
          "description": {
            "type": "string",
            "description": "할 일 상세 설명 (선택)"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "카테고리 ID (선택. 미지정 시 기본 카테고리 자동 배정 — BR-03)"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "시작일 (선택, BR-07)"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "종료일 (선택, BR-07). startDate 이후여야 함 (BR-06)"
          }
        },
        "required": ["title"]
      },
      "UpdateTodoRequest": {
        "type": "object",
        "description": "할 일 수정 요청 바디 (UC-06). 모든 필드 선택값이며 전송한 필드만 수정됨",
        "properties": {
          "title": {
            "type": "string",
            "description": "할 일 제목"
          },
          "description": {
            "type": "string",
            "description": "할 일 상세 설명"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "카테고리 ID"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "시작일 (BR-07)"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "종료일 (BR-07). startDate 이후여야 함 (BR-06)"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "표준 에러 응답 형식",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "에러 코드 (예: VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT)"
              },
              "message": {
                "type": "string",
                "description": "에러 메시지 (사람이 읽을 수 있는 설명)"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      }
    }
  },
  "paths": {
...
    "/api/todos": {
      "get": {
        "tags": ["할 일"],
        "summary": "할 일 목록 조회",
        "description": "현재 로그인한 사용자의 할 일 목록을 조회합니다. 쿼리 파라미터로 상태(BR-09~BR-12)와 카테고리(BR-08) 기반 필터링이 가능합니다. 타인의 할 일은 절대 포함되지 않습니다 (BR-02, UC-05).",
        "operationId": "getTodos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "할 일 상태 필터. pending(시작 전, BR-09), in_progress(진행 중, BR-10), overdue(기한 초과, BR-12), completed(완료, BR-11). 미입력 시 전체 조회",
            "schema": {
              "type": "string",
              "enum": ["pending", "in_progress", "overdue", "completed"]
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "description": "카테고리 ID 필터. 해당 카테고리에 속한 할 일만 반환 (BR-08)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "할 일 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                },
                "example": [
                  {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "userId": "550e8400-e29b-41d4-a716-446655440000",
                    "categoryId": "c1d2e3f4-a5b6-7890-abcd-ef1234567890",
                    "title": "프로젝트 기획서 작성",
                    "description": "Q3 로드맵 포함",
                    "startDate": "2026-05-28",
                    "endDate": "2026-06-01",
                    "isCompleted": false,
                    "createdAt": "2026-05-28T00:00:00.000Z",
                    "updatedAt": "2026-05-28T00:00:00.000Z"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "인증 토큰 없음 또는 만료 (BR-01)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다."
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_SERVER_ERROR",
                    "message": "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["할 일"],
        "summary": "할 일 등록",
        "description": "새로운 할 일을 등록합니다. 카테고리를 지정하지 않으면 '기본' 카테고리로 자동 배정됩니다 (BR-03). 종료일이 시작일보다 이전인 경우 400 오류가 반환됩니다 (BR-06). 시작일과 종료일은 필수값이 아닙니다 (BR-07, UC-04).",
        "operationId": "createTodo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              },
              "example": {
                "title": "프로젝트 기획서 작성",
                "description": "Q3 로드맵 포함",
                "categoryId": "c1d2e3f4-a5b6-7890-abcd-ef1234567890",
                "startDate": "2026-05-28",
                "endDate": "2026-06-01"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할 일 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "categoryId": "c1d2e3f4-a5b6-7890-abcd-ef1234567890",
                  "title": "프로젝트 기획서 작성",
                  "description": "Q3 로드맵 포함",
                  "startDate": "2026-05-28",
                  "endDate": "2026-06-01",
                  "isCompleted": false,
                  "createdAt": "2026-05-28T00:00:00.000Z",
                  "updatedAt": "2026-05-28T00:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 오류. 제목 누락 또는 종료일이 시작일보다 이전인 경우 (BR-06)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "종료일은 시작일보다 이전일 수 없습니다."
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 토큰 없음 또는 만료 (BR-01)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다."
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_SERVER_ERROR",
                    "message": "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "tags": ["할 일"],
        "summary": "할 일 단건 조회",
        "description": "특정 할 일의 상세 정보를 조회합니다. JWT에서 추출한 userId로 소유권을 검증하며, 본인 소유 할 일만 조회 가능합니다 (BR-02, BR-05).",
        "operationId": "getTodoById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "할 일 고유 식별자 (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "할 일 단건 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "categoryId": "c1d2e3f4-a5b6-7890-abcd-ef1234567890",
                  "title": "프로젝트 기획서 작성",
                  "description": "Q3 로드맵 포함",
                  "startDate": "2026-05-28",
                  "endDate": "2026-06-01",
                  "isCompleted": false,
                  "createdAt": "2026-05-28T00:00:00.000Z",
                  "updatedAt": "2026-05-28T00:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "인증 토큰 없음 또는 만료 (BR-01)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다."
                  }
                }
              }
            }
          },
          "403": {
            "description": "타인의 할 일 접근 시도. 소유권 검증 실패 (BR-02, BR-05)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "해당 리소스에 접근할 권한이 없습니다."
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 할 일 ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "해당 할 일을 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_SERVER_ERROR",
                    "message": "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["할 일"],
        "summary": "할 일 수정",
        "description": "특정 할 일의 내용을 수정합니다. 본인 소유 할 일만 수정 가능합니다 (BR-05). 종료일이 시작일보다 이전인 경우 400 오류가 반환됩니다 (BR-06). 전송한 필드만 수정되며 나머지는 기존 값이 유지됩니다 (UC-06).",
        "operationId": "updateTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "수정할 할 일의 고유 식별자 (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              },
              "example": {
                "title": "프로젝트 기획서 작성 (수정)",
                "endDate": "2026-06-05"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할 일 수정 성공. 수정된 할 일 정보 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "categoryId": "c1d2e3f4-a5b6-7890-abcd-ef1234567890",
                  "title": "프로젝트 기획서 작성 (수정)",
                  "description": "Q3 로드맵 포함",
                  "startDate": "2026-05-28",
                  "endDate": "2026-06-05",
                  "isCompleted": false,
                  "createdAt": "2026-05-28T00:00:00.000Z",
                  "updatedAt": "2026-05-28T12:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "종료일이 시작일보다 이전인 경우 (BR-06)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "종료일은 시작일보다 이전일 수 없습니다."
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 토큰 없음 또는 만료 (BR-01)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다."
                  }
                }
              }
            }
          },
          "403": {
            "description": "타인의 할 일 수정 시도. 소유권 검증 실패 (BR-02, BR-05)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "해당 리소스에 접근할 권한이 없습니다."
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 할 일 ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "해당 할 일을 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_SERVER_ERROR",
                    "message": "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["할 일"],
        "summary": "할 일 삭제",
        "description": "특정 할 일을 영구 삭제합니다 (복구 불가). 본인 소유 할 일만 삭제 가능합니다 (BR-05). 타인의 할 일 ID로 요청 시 403 오류가 반환됩니다 (UC-07).",
        "operationId": "deleteTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "삭제할 할 일의 고유 식별자 (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "할 일 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "삭제 완료 메시지"
                    }
                  }
                },
                "example": {
                  "message": "할 일이 삭제되었습니다."
                }
              }
            }
          },
          "401": {
            "description": "인증 토큰 없음 또는 만료 (BR-01)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다."
                  }
                }
              }
            }
          },
          "403": {
            "description": "타인의 할 일 삭제 시도. 소유권 검증 실패 (BR-02, BR-05)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "해당 리소스에 접근할 권한이 없습니다."
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 할 일 ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "해당 할 일을 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_SERVER_ERROR",
                    "message": "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{id}/complete": {
      "patch": {
        "tags": ["할 일"],
        "summary": "할 일 완료 처리",
        "description": "특정 할 일을 완료 상태로 변경합니다 (is_completed = true). 본인 소유 할 일만 완료 처리 가능합니다 (BR-05, UC-08).",
        "operationId": "completeTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "완료 처리할 할 일의 고유 식별자 (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "할 일 완료 처리 성공. 수정된 할 일 정보 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "userId": "550e8400-e29b-41d4-a716-446655440000",
                  "categoryId": "c1d2e3f4-a5b6-7890-abcd-ef1234567890",
                  "title": "프로젝트 기획서 작성",
                  "description": "Q3 로드맵 포함",
                  "startDate": "2026-05-28",
                  "endDate": "2026-06-01",
                  "isCompleted": true,
                  "createdAt": "2026-05-28T00:00:00.000Z",
                  "updatedAt": "2026-05-28T15:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "인증 토큰 없음 또는 만료 (BR-01)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다."
                  }
                }
              }
            }
          },
          "403": {
            "description": "타인의 할 일 완료 처리 시도. 소유권 검증 실패 (BR-02, BR-05)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "해당 리소스에 접근할 권한이 없습니다."
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 할 일 ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "해당 할 일을 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_SERVER_ERROR",
                    "message": "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}